requirements.txt
psycopg2-binary
MarkupSafe==2.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
pandas
requests
python-decouple==3.4